+++ src/Public/Test-WorkingDay.ps1
@@ -65,6 +65,9 @@ function Test-WorkingDay {
         [DateTime]$StartHour = '08:00:00',
 
         [Parameter()]
+        [ValidateScript({
+            if ($StartHour -gt $_) { throw "-StartHour must be less than -FinishHour." } else { return $true }
+        })]
         [DateTime]$FinishHour = '17:00:00',
 
         [Parameter()]