+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `Test-WorkingDay` now has parameter validation to ensure `-StartHour` is less then `-FinishHour`
+- `Get-ElapsedBusinessTime` now has parameter validation to ensure `-StartHour` is less then `-FinishHour`
 
 ## [0.1.6] - 2022-04-14
 ### Fixed